terminar crud produto

+++ projetoPHP/cms/cmsFaleConosco.php
@@ -2,24 +2,26 @@
 
     require_once ('./verificarUsuario.php');// VERIFICAR SE USUARIO ESTA LOGADO
 
-    //CONEXAO COM O BANCO
-    require_once('../bd/conexao.php');
-    $conexao = conexaoMySql();
+//    VERIFICAR SE O USUARIO LOGADO TEM PERMISSÃO PARA ACESSAR ESTA PÁGINA
+    if($_SESSION['adm_fale_conosco'] == 'ativado'){
+        //CONEXAO COM O BANCO
+        require_once('../bd/conexao.php');
+        $conexao = conexaoMySql();
 
-    //EXCLUIR REGISTRO CUJO CODIGO FOI PEGO NA URL
-    if(isset($_GET['codigo'])){
-        if($_GET['modo'] == 'excluir'){
-            $codMsg = $_GET['codigo'];
+        //EXCLUIR REGISTRO CUJO CODIGO FOI PEGO NA URL
+        if(isset($_GET['codigo'])){
+            if($_GET['modo'] == 'excluir'){
+                $codMsg = $_GET['codigo'];
 
-            $sql = "DELETE FROM tbl_fale_conosco WHERE cod_mensagem = ".$codMsg;
+                $sql = "DELETE FROM tbl_fale_conosco WHERE cod_mensagem = ".$codMsg;
 
-            if(mysqli_query($conexao, $sql)){
-                header('location: cmsFaleConosco.php');
-            }else{
-                echo("<script>alert('Erro ao excluir contato!');</script>");
+                if(mysqli_query($conexao, $sql)){
+                    header('location: cmsFaleConosco.php');
+                }else{
+                    echo("<script>alert('Erro ao excluir contato!');</script>");
+                }
             }
         }
-    }
 ?>
 
 <!DOCTYPE html>
@@ -31,13 +33,13 @@
         <meta charset="utf-8">
         <link rel="icon" href="../imgs/favicon.ico" type="image/x-icon">
         <script src="./js/jquery-3.3.1.min.js"></script>
-        <!-- FUNCAO DO AJAX DE ATIVA O MODAL DE VISUALIZAÇAO DO REGISTRO CLICADO -->
         <script>
             $(document).ready(function(){              
                 $('.visualizar').click(function(){
                     $('#container').fadeIn(300);
                 });
             });
+            // FUNCAO DO AJAX DE ATIVA O MODAL DE VISUALIZAÇAO DO REGISTRO CLICADO 
             function visualizarDados(codContato){
                 //↓ função do ajax para mandar informações para a modal de visualização de dados
                 $.ajax({
@@ -80,13 +82,16 @@
                         <tr class="table-titles">
                             <th id="title-nome">Nome</th>
                             <th id="title-email">E-mail</th>
-                            <th id="title-cel">Celular</th>
+                            <th id="title-cel">Assunto</th>
                             <th id="title-view">Visualizar</th>
                             <th class="title-excluir">Excluir</th>
                         </tr>
                         <?php
-                            // SCRIPT SQL QUE TRAZ OS REGISTROS DO BANCO
-                            $sql = "SELECT * FROM tbl_fale_conosco ORDER BY cod_mensagem DESC";
+                            // SCRIPT SQL QUE TRAZ OS REGISTROS DE MENSAGENS DE CLIENTES DO BANCO
+                            $sql = "SELECT * FROM tbl_fale_conosco
+                                    INNER JOIN tbl_assunto 
+                                    ON tbl_fale_conosco.cod_assunto = tbl_assunto.cod_assunto
+                                    ORDER BY tbl_fale_conosco.cod_assunto DESC";
 
                             $select = mysqli_query($conexao, $sql);
                             
@@ -95,7 +100,7 @@
                         <tr class="tables-registers">
                             <td class="txt-nome"><?php echo($rsContato['nome']); ?></td>
                             <td class="txt-email"><?php echo($rsContato['email']); ?></td>
-                            <td class="txt-cel"><?php echo($rsContato['celular']); ?></td>
+                            <td class="txt-assunto"><?php echo($rsContato['assunto']); ?></td>
                             <td class="txt-view">
                                 <figure>
                                     <!-- BOTAO QUE ATIVA MODAL PARA VISUALIZAÇAO DE REGISTRO -->
@@ -106,7 +111,7 @@
                                 <a href="?modo=excluir&codigo=<?php echo($rsContato['cod_mensagem']); ?>">
                                     <figure>
                                         <!-- BOTAO QUE EXCLUI O REGISTRO -->
-                                        <img onclick="return confirm('Deseja mesmo excluir <?php echo($rsContato['nome']); ?>?')" class="icon-delete" alt="Excluir Registro" title="Excluir Registro" src="icons/trash.png">
+                                        <img onclick="return confirm('Deseja excluir mensagem de <?php echo($rsContato['nome']); ?> permanentemente?')" class="icon-delete" alt="Excluir Registro" title="Excluir Registro" src="icons/trash.png">
                                     </figure>
                                 </a>
                             </td>
@@ -124,3 +129,13 @@
         </div>
     </body>
 </html>
+<?php
+    }else{
+        $userName = $_SESSION['user_name'];
+        echo
+        "<script>
+            alert('Usuário $userName não tem permissão de acesso à esta página.');
+            window.history.back();
+        </script>";
+    }
+?>
